Refactor : #113 scss에서 css in js 스타일로 변경
- styled-components를 설치하여 변경 진행중
- css in js로 body 프로퍼티 접근을 위해 createGlobalStyle을 사용
