+++ Client/src/components/Menu.jsx
@@ -6,6 +6,7 @@ const Menu = () => {
         <div>
             <ul>
                 <li><Link to="/">Home</Link></li>
+                <li><Link to="/login">Login</Link></li>
                 <li><Link to="/about">About</Link></li>
                 <li><Link to="/about/foo">About Foo</Link></li>
             </ul>
