--- Client/src/components/Routes.jsx
@@ -1,15 +1,0 @@
-import React from 'react';
-import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
-import LoginPage from '../pages/login/LoginPage.jsx';
-
-function Routes() {
-  return (
-    <Router>
-      <Switch>
-        <Route exact path="/" component={LoginPage} />
-      </Switch>
-    </Router>
-  );
-}
-
-export default Routes;
