+++ Client/src/pages/Home.jsx
@@ -1,8 +1,10 @@
 import React from 'react';
+import Menu from '../components/Menu'
 
 const Home = () => {
     return (
         <div>
+            <Menu></Menu>
             <h2>Home</h2>
         </div>
     );
