+++ Client/src/pages/index.jsx
@@ -1,2 +1,3 @@
 export { default as Home } from './Home';
 export { default as About } from './About';
+export { default as LoginPage } from './LoginPage';
