--- Client/src/pages/login/LoginPage.scss
@@ -1,68 +1,0 @@
-@import '../../scss/color';
-
-body {
-  background-color: $gray-100;
-}
-
-.container {
-  margin: auto;
-  margin-top: 100px;
-
-  .title {
-    text-align: center;
-  }
-
-  .login {
-    background-color: white;
-    border: solid 1px $gray-150;
-    border-radius: 5px;
-    font-weight: bold;
-    font-size: 14px;
-    padding: 10px 18px 22px 18px;
-
-    input {
-      border: solid 2px $gray-200;
-      border-radius: 3px;
-      box-shadow: 3px;
-      width: 230px;
-      height: 25px;
-    }
-
-    .routes {
-      display: flex;
-      justify-content: space-around;
-      margin-top: 10px;
-
-      a {
-        color: $blue-400;
-        text-decoration: none;
-      }
-    }
-
-    a {
-      text-decoration: none;
-
-      .github-login {
-        display: flex;
-        text-align: center;
-        background-color: $gray-400;
-        border: solid 1px $gray-450;
-        border-radius: 5px;
-        height: 35px;
-        margin-top: 20px;
-
-        p {
-          color: white;
-          font-weight: bold;
-          margin: auto;
-          margin-top: 8px;
-          margin-left: 40px;
-        }
-
-        img {
-          padding-right: 30px;
-        }
-      }
-    }
-  }
-}
