--- Client/src/shared/App.scss
@@ -1,8 +1,0 @@
-.App {
-  display: flex;
-}
-
-body {
-  margin: 0px;
-  padding: 0px;
-}
