--- README.md
@@ -1,11 +1,0 @@
-# IssueTracker-8
-
-## 모쨍이 팀원들
-- `J018 권예지` : 🐳
-- `J029 김도연` : 🐹
-- `J129 윤영우` : 🐕
-- `S012 김영렬` : 🐼
-- `S055 조수정` : 🐲
-
-## Sprint Backlog
-[Google Spreadsheet](https://docs.google.com/spreadsheets/d/19wkM--KlfBSZAe7_RBzZKZ5Rq0YNnLkuxhtWNhTGxDA/edit?usp=sharing)
